requirements.txt
allure-pytest-bdd==2.8.16
allure-python-commons==2.8.16
astroid==2.3.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
glob2==0.7
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
namedlist==1.7
packaging==20.1
parse==1.15.0
parse-type==0.5.2
pluggy==0.13.1
py==1.8.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
pytest-bdd==3.3.0
requests==2.23.0
selenium==3.141.0
seleniumwrapper==0.5.4
six==1.14.0
typed-ast==1.4.1
UnittestZero==0.2.0
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
zipp==2.1.0
webdriver-manager==3.2.0
pyautogui==0.9.52
python-dotenv==0.15.0
